Create AccuracyRecallPrecisionF1Score.py
To check the accurary of an alogrithm without depending on the SciKit learn Library.

Accuracy:
Accuracy measures how many classifications your algorithm got correct out of every classification it made   

Recall:
Accuracy can be an extremely misleading statistic depending on your data.
Recall measures the percentage of the relevant items your classifier was able to successfully find.

Precision:
Unfortunately, recall isn’t a perfect statistic either.
The statistic that will help demonstrate that this algorithm is flawed is precision.
Precision and recall are statistics that are on opposite ends of a scale. If one goes down, the other will go up.
Therefore, precision and recall are tied to each other.

F1 Score:
It is useful to consider the precision and recall of an algorithm, however, we still don’t have one number that can sufficiently describe how effective our algorithm is.
This is the job of the F1 score — F1 score is the harmonic mean of precision and recall. The harmonic mean of a group of numbers is a way to average them together. 
F1 score is a combination of precision and recall.
F1 score will be low if either precision or recall is low