Create PredictForLogisticRegression.py
In predict_class(), calculate the log-odds using the log_odds() function we defined earlier.
Still in predict_class(), find the probabilities that the samples belong to the positive class. Create a variable probabilities, and give it the value returned by calling sigmoid() on calculated_log_odds.
Return 1 for all values within probabilities equal to or above threshold, and 0 for all values below threshold.